api/agents/onboarding_chatbot/requirements.txt
crewai
crewai-tools
python-dotenv
# pysqlite3-binary